+++ services/kms/src/stackit/kms/api/default_api.py
@@ -1792,6 +1792,289 @@ class DefaultApi:
         )
 
     @validate_call
+    def delete_wrapping_key(
+        self,
+        project_id: Annotated[StrictStr, Field(description="The STACKIT portal project UUID the key ring is part of.")],
+        region_id: Annotated[StrictStr, Field(description="The STACKIT region name the key ring is located in.")],
+        key_ring_id: Annotated[StrictStr, Field(description="The key ring UUID.")],
+        wrapping_key_id: Annotated[StrictStr, Field(description="The wrapping key UUID.")],
+        _request_timeout: Union[
+            None,
+            Annotated[StrictFloat, Field(gt=0)],
+            Tuple[Annotated[StrictFloat, Field(gt=0)], Annotated[StrictFloat, Field(gt=0)]],
+        ] = None,
+        _request_auth: Optional[Dict[StrictStr, Any]] = None,
+        _content_type: Optional[StrictStr] = None,
+        _headers: Optional[Dict[StrictStr, Any]] = None,
+        _host_index: Annotated[StrictInt, Field(ge=0, le=0)] = 0,
+    ) -> None:
+        """Delete wrapping key
+
+        Deletes the given wrapping key
+
+        :param project_id: The STACKIT portal project UUID the key ring is part of. (required)
+        :type project_id: str
+        :param region_id: The STACKIT region name the key ring is located in. (required)
+        :type region_id: str
+        :param key_ring_id: The key ring UUID. (required)
+        :type key_ring_id: str
+        :param wrapping_key_id: The wrapping key UUID. (required)
+        :type wrapping_key_id: str
+        :param _request_timeout: timeout setting for this request. If one
+                                 number provided, it will be total request
+                                 timeout. It can also be a pair (tuple) of
+                                 (connection, read) timeouts.
+        :type _request_timeout: int, tuple(int, int), optional
+        :param _request_auth: set to override the auth_settings for an a single
+                              request; this effectively ignores the
+                              authentication in the spec for a single request.
+        :type _request_auth: dict, optional
+        :param _content_type: force content-type for the request.
+        :type _content_type: str, Optional
+        :param _headers: set to override the headers for a single
+                         request; this effectively ignores the headers
+                         in the spec for a single request.
+        :type _headers: dict, optional
+        :param _host_index: set to override the host_index for a single
+                            request; this effectively ignores the host_index
+                            in the spec for a single request.
+        :type _host_index: int, optional
+        :return: Returns the result object.
+        """  # noqa: E501 docstring might be too long
+
+        _param = self._delete_wrapping_key_serialize(
+            project_id=project_id,
+            region_id=region_id,
+            key_ring_id=key_ring_id,
+            wrapping_key_id=wrapping_key_id,
+            _request_auth=_request_auth,
+            _content_type=_content_type,
+            _headers=_headers,
+            _host_index=_host_index,
+        )
+
+        _response_types_map: Dict[str, Optional[str]] = {
+            "204": None,
+            "400": "HttpError",
+            "401": "HttpError",
+            "404": "HttpError",
+            "500": "HttpError",
+        }
+        response_data = self.api_client.call_api(*_param, _request_timeout=_request_timeout)
+        response_data.read()
+        return self.api_client.response_deserialize(
+            response_data=response_data,
+            response_types_map=_response_types_map,
+        ).data
+
+    @validate_call
+    def delete_wrapping_key_with_http_info(
+        self,
+        project_id: Annotated[StrictStr, Field(description="The STACKIT portal project UUID the key ring is part of.")],
+        region_id: Annotated[StrictStr, Field(description="The STACKIT region name the key ring is located in.")],
+        key_ring_id: Annotated[StrictStr, Field(description="The key ring UUID.")],
+        wrapping_key_id: Annotated[StrictStr, Field(description="The wrapping key UUID.")],
+        _request_timeout: Union[
+            None,
+            Annotated[StrictFloat, Field(gt=0)],
+            Tuple[Annotated[StrictFloat, Field(gt=0)], Annotated[StrictFloat, Field(gt=0)]],
+        ] = None,
+        _request_auth: Optional[Dict[StrictStr, Any]] = None,
+        _content_type: Optional[StrictStr] = None,
+        _headers: Optional[Dict[StrictStr, Any]] = None,
+        _host_index: Annotated[StrictInt, Field(ge=0, le=0)] = 0,
+    ) -> ApiResponse[None]:
+        """Delete wrapping key
+
+        Deletes the given wrapping key
+
+        :param project_id: The STACKIT portal project UUID the key ring is part of. (required)
+        :type project_id: str
+        :param region_id: The STACKIT region name the key ring is located in. (required)
+        :type region_id: str
+        :param key_ring_id: The key ring UUID. (required)
+        :type key_ring_id: str
+        :param wrapping_key_id: The wrapping key UUID. (required)
+        :type wrapping_key_id: str
+        :param _request_timeout: timeout setting for this request. If one
+                                 number provided, it will be total request
+                                 timeout. It can also be a pair (tuple) of
+                                 (connection, read) timeouts.
+        :type _request_timeout: int, tuple(int, int), optional
+        :param _request_auth: set to override the auth_settings for an a single
+                              request; this effectively ignores the
+                              authentication in the spec for a single request.
+        :type _request_auth: dict, optional
+        :param _content_type: force content-type for the request.
+        :type _content_type: str, Optional
+        :param _headers: set to override the headers for a single
+                         request; this effectively ignores the headers
+                         in the spec for a single request.
+        :type _headers: dict, optional
+        :param _host_index: set to override the host_index for a single
+                            request; this effectively ignores the host_index
+                            in the spec for a single request.
+        :type _host_index: int, optional
+        :return: Returns the result object.
+        """  # noqa: E501 docstring might be too long
+
+        _param = self._delete_wrapping_key_serialize(
+            project_id=project_id,
+            region_id=region_id,
+            key_ring_id=key_ring_id,
+            wrapping_key_id=wrapping_key_id,
+            _request_auth=_request_auth,
+            _content_type=_content_type,
+            _headers=_headers,
+            _host_index=_host_index,
+        )
+
+        _response_types_map: Dict[str, Optional[str]] = {
+            "204": None,
+            "400": "HttpError",
+            "401": "HttpError",
+            "404": "HttpError",
+            "500": "HttpError",
+        }
+        response_data = self.api_client.call_api(*_param, _request_timeout=_request_timeout)
+        response_data.read()
+        return self.api_client.response_deserialize(
+            response_data=response_data,
+            response_types_map=_response_types_map,
+        )
+
+    @validate_call
+    def delete_wrapping_key_without_preload_content(
+        self,
+        project_id: Annotated[StrictStr, Field(description="The STACKIT portal project UUID the key ring is part of.")],
+        region_id: Annotated[StrictStr, Field(description="The STACKIT region name the key ring is located in.")],
+        key_ring_id: Annotated[StrictStr, Field(description="The key ring UUID.")],
+        wrapping_key_id: Annotated[StrictStr, Field(description="The wrapping key UUID.")],
+        _request_timeout: Union[
+            None,
+            Annotated[StrictFloat, Field(gt=0)],
+            Tuple[Annotated[StrictFloat, Field(gt=0)], Annotated[StrictFloat, Field(gt=0)]],
+        ] = None,
+        _request_auth: Optional[Dict[StrictStr, Any]] = None,
+        _content_type: Optional[StrictStr] = None,
+        _headers: Optional[Dict[StrictStr, Any]] = None,
+        _host_index: Annotated[StrictInt, Field(ge=0, le=0)] = 0,
+    ) -> RESTResponseType:
+        """Delete wrapping key
+
+        Deletes the given wrapping key
+
+        :param project_id: The STACKIT portal project UUID the key ring is part of. (required)
+        :type project_id: str
+        :param region_id: The STACKIT region name the key ring is located in. (required)
+        :type region_id: str
+        :param key_ring_id: The key ring UUID. (required)
+        :type key_ring_id: str
+        :param wrapping_key_id: The wrapping key UUID. (required)
+        :type wrapping_key_id: str
+        :param _request_timeout: timeout setting for this request. If one
+                                 number provided, it will be total request
+                                 timeout. It can also be a pair (tuple) of
+                                 (connection, read) timeouts.
+        :type _request_timeout: int, tuple(int, int), optional
+        :param _request_auth: set to override the auth_settings for an a single
+                              request; this effectively ignores the
+                              authentication in the spec for a single request.
+        :type _request_auth: dict, optional
+        :param _content_type: force content-type for the request.
+        :type _content_type: str, Optional
+        :param _headers: set to override the headers for a single
+                         request; this effectively ignores the headers
+                         in the spec for a single request.
+        :type _headers: dict, optional
+        :param _host_index: set to override the host_index for a single
+                            request; this effectively ignores the host_index
+                            in the spec for a single request.
+        :type _host_index: int, optional
+        :return: Returns the result object.
+        """  # noqa: E501 docstring might be too long
+
+        _param = self._delete_wrapping_key_serialize(
+            project_id=project_id,
+            region_id=region_id,
+            key_ring_id=key_ring_id,
+            wrapping_key_id=wrapping_key_id,
+            _request_auth=_request_auth,
+            _content_type=_content_type,
+            _headers=_headers,
+            _host_index=_host_index,
+        )
+
+        _response_types_map: Dict[str, Optional[str]] = {
+            "204": None,
+            "400": "HttpError",
+            "401": "HttpError",
+            "404": "HttpError",
+            "500": "HttpError",
+        }
+        response_data = self.api_client.call_api(*_param, _request_timeout=_request_timeout)
+        return response_data.response
+
+    def _delete_wrapping_key_serialize(
+        self,
+        project_id,
+        region_id,
+        key_ring_id,
+        wrapping_key_id,
+        _request_auth,
+        _content_type,
+        _headers,
+        _host_index,
+    ) -> RequestSerialized:
+
+        _host = None
+
+        _collection_formats: Dict[str, str] = {}
+
+        _path_params: Dict[str, str] = {}
+        _query_params: List[Tuple[str, str]] = []
+        _header_params: Dict[str, Optional[str]] = _headers or {}
+        _form_params: List[Tuple[str, str]] = []
+        _files: Dict[str, Union[str, bytes]] = {}
+        _body_params: Optional[bytes] = None
+
+        # process the path parameters
+        if project_id is not None:
+            _path_params["projectId"] = project_id
+        if region_id is not None:
+            _path_params["regionId"] = region_id
+        if key_ring_id is not None:
+            _path_params["keyRingId"] = key_ring_id
+        if wrapping_key_id is not None:
+            _path_params["wrappingKeyId"] = wrapping_key_id
+        # process the query parameters
+        # process the header parameters
+        # process the form parameters
+        # process the body parameter
+
+        # set the HTTP header `Accept`
+        if "Accept" not in _header_params:
+            _header_params["Accept"] = self.api_client.select_header_accept(["application/json"])
+
+        # authentication setting
+        _auth_settings: List[str] = []
+
+        return self.api_client.param_serialize(
+            method="DELETE",
+            resource_path="/v1beta/projects/{projectId}/regions/{regionId}/keyrings/{keyRingId}/wrappingkeys/{wrappingKeyId}",
+            path_params=_path_params,
+            query_params=_query_params,
+            header_params=_header_params,
+            body=_body_params,
+            post_params=_form_params,
+            files=_files,
+            auth_settings=_auth_settings,
+            collection_formats=_collection_formats,
+            _host=_host,
+            _request_auth=_request_auth,
+        )
+
+    @validate_call
     def destroy_version(
         self,
         project_id: Annotated[StrictStr, Field(description="The STACKIT portal project UUID the key ring is part of.")],