+++ services/kms/src/stackit/kms/models/key.py
@@ -80,8 +80,10 @@ class Key(BaseModel):
     @field_validator("state")
     def state_validate_enum(cls, value):
         """Validates the enum"""
-        if value not in set(["active", "version_not_ready", "deleted"]):
-            raise ValueError("must be one of enum values ('active', 'version_not_ready', 'deleted')")
+        if value not in set(["active", "deleted", "not_available", "errors_exist", "no_version"]):
+            raise ValueError(
+                "must be one of enum values ('active', 'deleted', 'not_available', 'errors_exist', 'no_version')"
+            )
         return value
 
     model_config = ConfigDict(