+++ services/kms/src/stackit/kms/models/wrapping_key.py
@@ -69,8 +69,10 @@ class WrappingKey(BaseModel):
     @field_validator("state")
     def state_validate_enum(cls, value):
         """Validates the enum"""
-        if value not in set(["active", "key_material_not_ready", "expired", "deleting"]):
-            raise ValueError("must be one of enum values ('active', 'key_material_not_ready', 'expired', 'deleting')")
+        if value not in set(["active", "key_material_not_ready", "expired", "deleted", "key_material_unavailable"]):
+            raise ValueError(
+                "must be one of enum values ('active', 'key_material_not_ready', 'expired', 'deleted', 'key_material_unavailable')"
+            )
         return value
 
     model_config = ConfigDict(